+++ app/build.gradle
@@ -13,7 +13,7 @@ android {
 }
 
 dependencies {
-    implementation(projects.common.ui)
+    implementation(projects.core.ui)
     implementation(projects.camerax)
     implementation(projects.legacy)
     implementation(projects.slide)
