+++ build.gradle
@@ -1,4 +1,4 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
+// Top-level build file where you can add configuration options core to all sub-projects/modules.
 
 
 buildscript {
