+++ camerax/build.gradle
@@ -13,7 +13,7 @@ android {
 }
 
 dependencies {
-    implementation(projects.common.ui)
+    implementation(projects.core.ui)
     implementation(libs.androidx.compose.foundation)
     implementation(libs.androidx.compose.material)
     debugImplementation(libs.androidx.compose.ui.tooling)
