+++ core/ui/proguard-rules.pro
