+++ core/ui/src/main/java/net/kwmt27/jetpackcomposeplayground/core/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package net.kwmt27.jetpackcomposeplayground.common.ui.theme
+package net.kwmt27.jetpackcomposeplayground.core.ui.theme
 
 import androidx.compose.ui.graphics.Color
