+++ legacy/build.gradle.kts
@@ -8,7 +8,7 @@ android {
 }
 
 dependencies {
-    implementation(projects.common.ui)
+    implementation(projects.core.ui)
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.google.material)
